# specs/001-primoinspect-digital-inspection/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: PrimoInspect Digital Inspection Platform API
  description: REST API for renewable energy inspection management with mobile-first design
  version: 1.0.0
  contact:
    name: PrimoInspect API Support
    email: api@primoinspect.com
servers:
  - url: https://api.primoinspect.com/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /projects:
    get:
      summary: List projects for authenticated user
      tags: [Projects]
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, on-hold, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create new project (Project Manager only)
      tags: [Projects]
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        description:
          type: string
        status:
          type: string
          enum: [active, completed, on-hold, cancelled]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          $ref: "#/components/schemas/Location"
        teamMembers:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 10
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProject:
      type: object
      required: [name, startDate, location]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          $ref: "#/components/schemas/Location"
...
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracy:
          type: number
          minimum: 0
        address:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...